+++ PureInvoke/Functions/Invoke-AdvApiLsaClose.ps1
@@ -1,0 +1,31 @@
+
+function Invoke-AdvApiLsaClose
+{
+    <#
+    .SYNOPSIS
+    Calls the advapi32.dll library's `LsaClose` method to close an LSA policy handle.
+
+    .DESCRIPTION
+    The `Invoke-AdvApiLsaClose` function calls the advapi32.dll library's `LsaClose` method to close an LSA policy
+    handle that was created with `Invoke-AdvApiLsaOpenPolicy`. Pass the policy handle to the `PolicyHandle` parameter.
+    The function closes the policy and returns `$true` if the close succeeded. If the close fails, returns `$false` and
+    writes an error.
+
+    .EXAMPLE
+    Invoke-AdvApiLsaClose -PolicyHandle $handle
+
+    Demonstrates how to call `Invoke-AdvApiLsaClose`.
+    #>
+    [CmdletBinding()]
+    param(
+        # The policy handle to close. Use `Invoke-AdvApiLsaOpenPolicy` to create policy handles.
+        [Parameter(Mandatory)]
+        [IntPtr] $PolicyHandle
+    )
+
+    Set-StrictMode -Version 'Latest'
+    Use-CallerPreference -Cmdlet $PSCmdlet -Session $ExecutionContext.SessionState
+
+    $ntstatus = [PureInvoke.AdvApi32]::LsaClose($PolicyHandle)
+    Assert-NTStatusSuccess -Status $ntstatus -Message 'Invoke-AdvApiLsaClose failed'
+}