+++ PureInvoke/Functions/Write-Win32Error.ps1
@@ -4,7 +4,7 @@ function Write-Win32Error
     [CmdletBinding()]
     param(
         [Parameter(Mandatory)]
-        [PureInvoke.ErrorCode] $ErrorCode,
+        [int] $ErrorCode,
 
         [String] $Message
     )
@@ -14,11 +14,18 @@ function Write-Win32Error
 
     if ($Message)
     {
-        $Message.TrimEnd('.')
+        $Message = $Message.TrimEnd('.')
         $Message = "${Message}: "
     }
 
     $win32Ex = [Win32Exception]::New($ErrorCode)
-    $msg = "${Message}$($win32Ex.Message) (0x$($win32Ex.ErrorCode.ToString('x'))/$($win32Ex.NativeErrorCode))."
+
+    $period = '.'
+    if ($win32ex.Message.EndsWith('.'))
+    {
+        $period = ''
+    }
+
+    $msg = "${Message}$($win32Ex.Message)${period} (0x$($win32Ex.ErrorCode.ToString('x'))/$($win32Ex.NativeErrorCode))"
     Write-Error -Message $msg -ErrorAction $ErrorActionPreference
 }